+++ .gitignore
@@ -74,3 +74,6 @@ typings/
 
 # FuseBox cache
 .fusebox/
+
+#build
+dist/
