+++ package.json
@@ -2,24 +2,27 @@
   "name": "spike-bot",
   "version": "0.3.0",
   "description": "Discord Bot for discord based projects",
-  "main": "index.ts",
+  "main": "dist/bot.js",
   "scripts": {
-    "start": "npx ts-node ./index.ts",
-    "dev": " npx nodemon ./index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "ts-node src/bot.ts",
+    "dev": "nodemon src/bot.ts --no-deployment",
+    "build": "tsc && cp .env dist/"
   },
-  "author": "John William Bowen",
+  "author": {
+    "name": "John William Bowen",
+    "url": "https://github.com/bowenjw",
+    "email": "johnwb1776@gmail.com"
+  },
   "license": "GPL-3.0-or-later",
   "dependencies": {
-    "@discordjs/rest": "^1.4.0",
     "discord.js": "^14.7.1",
     "dotenv": "^16.0.3",
-    "mongoose": "^6.8.0",
+    "mongoose": "^6.8.1",
     "ts-node": "^10.9.1"
   },
   "devDependencies": {
-    "discord-api-types": "^0.37.21",
-    "eslint": "^8.29.0",
+    "@types/node": "^18.11.17",
+    "eslint": "^8.30.0",
     "nodemon": "^2.0.20",
     "typescript": "^4.9.4"
   }
