--- src/classes/config.json
@@ -1,14 +1,0 @@
-{
-    "guild": "1038425229300613150",
-    "interactions": {
-        "receiveMessageComponents": true,
-        "receiveModals": true,
-        "replyOnError": true,
-        "splitCustomId": false,
-        "useGuildCommands": true
-    },
-    "colors": {
-        "embed": "#2f3136"
-    },
-    "restVersion": "10"
-}
