+++ src/commands/chat/ping.ts
@@ -1,13 +1,14 @@
 import { ChatInputCommandInteraction, EmbedBuilder, PermissionFlagsBits, SlashCommandBuilder } from 'discord.js';
-import { client } from '../../index';
+import { client } from '../../bot';
 
-export const slashCommandBuilder = new SlashCommandBuilder()
+export const builder = new SlashCommandBuilder()
 	.setName('ping')
 	.setDescription('Gets the current latencey of the bot')
     .setDefaultMemberPermissions(PermissionFlagsBits.SendMessages)
-	.setDMPermission(true)
+	.setDMPermission(true),
+global = true
 
-export async function commandExecute(interaction: ChatInputCommandInteraction) {
+export async function execute(interaction: ChatInputCommandInteraction) {
 	const embed = new EmbedBuilder()
 		.setColor("Blurple")
 		.setTitle('Current Latencey')
