+++ src/features/index.ts
@@ -1,0 +1,3 @@
+export { deploy} from './deploy'
+export { amputator} from './amputator'
+export { timeoutEmbed } from './timeout'
