+++ src/interfaces/GuildSchema.ts
@@ -1,0 +1,16 @@
+import { Snowflake } from "discord.js"
+
+export interface ISystem {
+    channel?: Snowflake,
+    enabled: boolean
+}
+export interface IwarnningSystem extends ISystem {
+    appealMessage?: string,
+    maxActiveWarns: number
+}
+export interface IGuild{
+    id: Snowflake,
+    name: string,
+    warning: IwarnningSystem
+    timeoutlog: ISystem
+}
