+++ src/interfaces/index.ts
@@ -1,3 +1,6 @@
-export { Command } from './Command';
-export { Event } from './Event';
-export { Interaction } from './Interaction';
+export { Command as Icommand} from './Command';
+export { Event as Ievent } from './Event';
+export { Iinteraction } from './Interaction';
+export { IGuild, ISystem, IwarnningSystem } from './GuildSchema'
+export { Iwarn } from './WarningSchema'
+export { AmputatorResponse } from './Amputator'
