+++ src/schema/index.ts
@@ -1,0 +1,2 @@
+export { guilds } from './guilds'
+export { warnings } from './warnings'
